Refactor process2 data structure
