+++ modules/process2/controllers/TaskDataTestController.php
@@ -4,9 +4,9 @@ namespace app\modules\process2\controllers;
 
 use app\controllers\BaseController;
 use app\modules\process\models\task_data\Req3TasksDataItems;
-use app\modules\process2\data\widget\IdentifierViewTypeWidget;
-use app\modules\process2\data\factory\DataItemDtoFactory;
-use app\modules\process2\data\service\loader\DataItemBatchLoader;
+use app\modules\process2\data\widgets\view\IdentifierViewTypeWidget;
+use app\modules\process2\data\factories\DataItemDtoFactory;
+use app\modules\process2\data\services\loader\DataItemBatchLoader;
 use Yii;
 use yii\web\Controller;
