+++ modules/process2/data/services/DataItemIdentifierRegistry.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace app\modules\process2\data\service;
+namespace app\modules\process2\data\services;
 
 use app\modules\process2\identifier\BaseIdentifier;
 use app\modules\process2\identifier\preset\map\IdentifierMapProvider;
 use app\modules\process2\data\dto\DataItemDto;
-use app\modules\process2\data\service\loader\DataItemLoaderInterface;
+use app\modules\process2\data\services\loader\DataItemLoaderInterface;
 use Yii;
