+++ modules/process2/data/services/loader/DataItemLoaderInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace app\modules\process2\data\service\loader;
+namespace app\modules\process2\data\services\loader;
 
 use app\modules\process2\data\dto\DataItemDto;
