+++ modules/process2/data/services/loader/type/ServiceDataItemLoader.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace app\modules\process2\data\service\loader\type;
+namespace app\modules\process2\data\services\loader\type;
 
 use app\modules\process2\data\dto\type\DataItemServiceDto;
-use app\modules\process2\data\service\loader\DataItemLoaderInterface;
+use app\modules\process2\data\services\loader\DataItemLoaderInterface;
 use app\modules\utm\models\ServicesData;
 use yii\helpers\ArrayHelper;
