+++ modules/process2/data/services/loader/type/UserDataItemLoader.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace app\modules\process2\data\service\loader\type;
+namespace app\modules\process2\data\services\loader\type;
 
 use app\modules\process2\data\dto\type\DataItemUserDto;
-use app\modules\process2\data\service\loader\DataItemLoaderInterface;
+use app\modules\process2\data\services\loader\DataItemLoaderInterface;
 use app\modules\user\models\Users;
 use yii\helpers\ArrayHelper;
