+++ modules/process2/data/services/loader/type/WarehouseDataItemLoader.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace app\modules\process2\data\service\loader\type;
+namespace app\modules\process2\data\services\loader\type;
 
 use app\modules\process2\data\dto\type\DataItemWarehouseDto;
-use app\modules\process2\data\service\loader\DataItemLoaderInterface;
+use app\modules\process2\data\services\loader\DataItemLoaderInterface;
 use app\modules\warehouse\models\Warehouse;
 use yii\helpers\ArrayHelper;
