+++ modules/process2/data/widgets/edit/BaseIdentifierInputWidget.php
@@ -1,5 +1,5 @@
 <?php
-namespace app\modules\process2\data\widget;
+namespace app\modules\process2\data\widgets\edit;
 
 use app\modules\process2\data\dto\DataItemDto;
 use yii\base\Widget;
