+++ modules/process2/data/widgets/view/BaseIdentifierViewWidget.php
@@ -1,5 +1,5 @@
 <?php
-namespace app\modules\process2\data\widget;
+namespace app\modules\process2\data\widgets\view;
 
 use app\modules\process2\data\dto\DataItemDto;
 use yii\base\Widget;
