+++ modules/process2/data/widgets/view/types/ServiceViewWidget.php
@@ -1,8 +1,8 @@
 <?php
-namespace app\modules\process2\data\widget\type\view;
+namespace app\modules\process2\data\widgets\view\types;
 
 use app\modules\process2\data\dto\type\DataItemServiceDto;
-use app\modules\process2\data\widget\BaseIdentifierViewWidget;
+use app\modules\process2\data\widgets\view\BaseIdentifierViewWidget;
 
 final class ServiceViewWidget extends BaseIdentifierViewWidget
 {
